functions/instance_scheduler/requirements.txt
google-cloud-datastore>=2.1.0
google-cloud-compute>=0.2.1